More Array Functions
Our task is to implement functions that enable us to perform some arithmetic operations on
positive integer values that have not limit to their scope. In the basic form, the operation that
we are interested in is addition.
In order to enable us to bypass the range of integer values, we will represent an integer
number as an array of integer digits.
The first functions are for the conversion of numbers to that mentioned above array. The size
of the array is to be adjusted to the parameter value.
int [] toNumberArray(int value)
e.g. toNumberArray(123) should return an array: [1,2,3]
int [] toNumberArray(String value)
e.g. toNumberArray("7634") should return an array[7, 6,3, 4]. If the string value does not
represent an integer number then the function should return an empty array.
As stated above, the basic supported arithmetic operation is addition.
int [] addNumberArrays(int[] a1, int[] a2)
for:
int [] a1={1, 9}
int [] a2={8, 2, 3}
e.g. for function call: addNumberArrays(a1, a2} we should get [8, 4, 2] as a result.
In order to perform the addition, you may use the method you have learnt as a pupil in your
elementary school.
For those interested, here are the possible extensions:
1. Taking care of negative numbers
2. Implementing a method for multiplication